% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_forest.R
\name{add_forest}
\alias{add_forest}
\title{Add a Forest Plot}
\usage{
add_forest(x)
}
\arguments{
\item{x}{a gtsummary table}
}
\value{
a gt table
}
\description{
Add a Forest Plot
}
\examples{
# Add example
}
