% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtforester-package.R
\docType{package}
\name{gtforester-package}
\alias{gtforester}
\alias{gtforester-package}
\title{gtforester: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [copyright holder]

}
\keyword{internal}
